#44 [2h] extract room members ui and fix owner role display
