+++ src/pages/Devices.css
@@ -40,137 +40,6 @@
   margin-top: 24px;
 }
 
-.members-section {
-  margin-top: 2rem;
-  padding: 1.25rem;
-  border: 1px solid #e5e7eb;
-  border-radius: 16px;
-  background: #fff;
-}
-
-.members-section-header h3,
-.members-section-header p,
-.member-card p {
-  margin: 0;
-}
-
-.members-section-header p {
-  margin-top: 0.35rem;
-  color: #6b7280;
-}
-
-.member-invite-form {
-  display: flex;
-  gap: 0.75rem;
-  margin-top: 1rem;
-}
-
-.member-invite-form input {
-  flex: 1;
-  border: 1px solid #d1d5db;
-  border-radius: 10px;
-  padding: 0.75rem 0.9rem;
-  font: inherit;
-}
-
-.member-invite-form button {
-  min-width: 3rem;
-  border: none;
-  border-radius: 10px;
-  background: var(--primary);
-  color: #fff;
-  font-size: 1.2rem;
-  cursor: pointer;
-}
-
-.member-invite-form button:disabled {
-  cursor: not-allowed;
-  opacity: 0.6;
-}
-
-.members-error,
-.members-success {
-  margin: 0.9rem 0 0;
-  padding: 0.75rem 0.9rem;
-  border-radius: 10px;
-}
-
-.members-error {
-  background: #fef2f2;
-  color: #b91c1c;
-}
-
-.members-success {
-  background: #ecfdf5;
-  color: #047857;
-}
-
-.member-list {
-  display: grid;
-  gap: 0.9rem;
-  margin-top: 1rem;
-}
-
-.member-card {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 1rem;
-  padding: 1rem;
-  border-radius: 12px;
-  background: #f8fafc;
-  border: 1px solid #e5e7eb;
-}
-
-.member-card-pending {
-  background: #fff7ed;
-  border-color: #fdba74;
-}
-
-.member-card-declined {
-  background: #fef2f2;
-  border-color: #fca5a5;
-}
-
-.member-email {
-  font-weight: 600;
-}
-
-.member-role {
-  margin-top: 0.3rem !important;
-  color: #6b7280;
-  text-transform: capitalize;
-}
-
-.member-status {
-  margin-top: 0.3rem !important;
-  color: #c2410c;
-  font-weight: 600;
-}
-
-.member-invite-actions {
-  display: flex;
-  gap: 0.75rem;
-}
-
-.member-resend-button {
-  border: none;
-  border-radius: 10px;
-  padding: 0.65rem 0.9rem;
-  background: #dbeafe;
-  color: #1d4ed8;
-  cursor: pointer;
-}
-
-.member-remove-button {
-  border: none;
-  border-radius: 10px;
-  padding: 0.65rem 0.9rem;
-  background: #fef2f2;
-  color: #b91c1c;
-  cursor: pointer;
-}
-
 /* --- Mobile --- */
 @media (max-width: 900px) {
   .devices-layout {
@@ -210,16 +79,6 @@
     justify-content: center;
     grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
     gap: 16px;
-  }
-
-  .member-card,
-  .member-invite-form {
-    flex-direction: column;
-    align-items: stretch;
-  }
-
-  .member-invite-actions {
-    flex-direction: column;
   }
 }
